linux: Address warnings about missing initializers
It looks like gcc (GCC) 13.2.1 20231011 (Red Hat 13.2.1-4) thinks it's a good
idea to complain about every single zero-initialized C structure.

This is widely acknowledged as a (rather annoying)
[GCC bug](https://gcc.gnu.org/bugzilla/show_bug.cgi?id=53119)

Add a constructor to silence the warning where possible.
For DMCP C structures where this is not possible, well, add ugly code.