ui: Emulate HP48/HP50G behavior for errors
When there is an error, DB48X used to require the error to be cleared
explicitly using the EXIT, BSP or ENTER key.

The HP48 and HP50 do not behave that way. Instead, they accept key
entries when an error message is shown, and clear the error.

Make the HP behavior the default, but with a flag making it possible
to restore the previous behaviour, `NeedToClearErrors`.

Signed-off-by: Christophe de Dinechin <christophe@dinechin.org>