lists: Separate list sum/product from regular sum/product
For some reason, I used the `Sum` (`Σ`) and `Product` (`∏`) commands
for list sums and products. This is not the way the HP48/HP50 do it,
so it's better to have dedicated `ΣList` and `∏List` commands.

Also implemented the missing `∆List` command and added the missing
related tests for all these commands.

Signed-off-by: Christophe de Dinechin <christophe@dinechin.org>