command: Factor out arity for all commands
For all commands, put the arity check in the top-level evaluator.
This factors out all that code in a single (source-code) location.

This also makes it possible to use arity reliably during parsing of
expressions, and fixes a number of cases where it wqas set wrong, like
for the `det` command and many others.

Finally, it fixes missing calls to `rt.args` for a number of commands,
while explicitly removing it (`nargs < 0`) for a few cases where it
was harful, like `LastArgs` or `Undo`.

Fixes: #865

Signed-off-by: Christophe de Dinechin <christophe@dinechin.org>