graphics: Fix bug drawing a line of width 0
When the line width is set to 0, we were incorrectly taking
`(width-1)/2` for the negative offset, resulting in a huge box being
drawn instead of a thin line.

Arguably, a line width of 0 should maybe not be drawn at all?
But this makes it more likely that someone will run the test, see a
blank screen and think this is buggy.

Fixes: #880

Signed-off-by: Christophe de Dinechin <christophe@dinechin.org>