files: Open only one configuration file at a time
When parsing the value of a constant with a unit, close the outer
constant configuration file before parsing the constants body (because
parsing itself does require opening another configuration file).

Similar scenarios occur with character configuration file while
building the character menus, and with units when we evaluate the body
of a unit.

Fixes: #882

Signed-off-by: Christophe de Dinechin <christophe@dinechin.org>