functions: Turn `min` and `max` into functions
The `min` and `max` commands were declared as arithmetic operators
instead of n-ary function. This had a number of bad effects, including
the fact that no parentheses were inserted when entering them using a
menu entry, but also the lack of proper handling for the symbolic
case.

Turn them into n-ary functions, which requires adding support for the
text type when processing arguments.

Fixes: #896

Signed-off-by: Christophe de Dinechin <christophe@dinechin.org>