rewrites: Enhance `↓Match` and `↑Match` to return number of changes
The result of `↓Match` and `↑Match` on HP calculators is `0` for no
rewrites, or `1` if a replacement occured.

However, it's possible (rarely) for more than one rewrite to occur
even on HP (it's difficult because the algorithm appears to be
single-pass). DB48X can perform multiple rewrites on the same
expression, so it makes sense to enhance `↓Match` and `↑Match` to
return the number of replacements performed. This can still be used as
a test.

Also document the differences in terms of pattern.

Fixes: #912

Signed-off-by: Christophe de Dinechin <christophe@dinechin.org>