polynomial: Add polynomial division
Implement polynomial euclidean division, which gives a meaning to
divide and remainder for polynomials.

Polynomial division is only well-defined with a single variable.
On HP calculators, this is configured by the `VX` variable in the
`CASDIR` directory. A similar mechanism is provided allowing the
creation of an algebraic configuration directory.

Signed-off-by: Christophe de Dinechin <christophe@dinechin.org>