symbols: Avoid ignoring the list of bad characters in symbols
On DM32/DM42, the test `isalnum(cp)` apparently returns `true` for
some values that are outside of the ASCII range. This causes
`is_valid_in_name()` to skip the test for invalid Unicode characters,
and notably allows `⁻` in a name. So `x⁻` was parsed as a valid
symbol, which caused `x⁻¹` to fail to parse.

As a result, all the constants with `⁻¹` in the unit definition would
fail to parse, and we could not get teh numerical value.

Fixes: #920

Signed-off-by: Christophe de Dinechin <christophe@dinechin.org>