solver: Implement solving menu
The `SolvingMenu` makes it possible to quickly solve equations with
multiple variables. It is typically accessed from the `SolverMenu`
(currently under key F4).

The menu shows the variables in the current equation, as defined by
the `'EQ'` variable (aka `Equation`). This can be accessed with the
`StEQ` and `RcEQ` commands.

For each variable in the equation, the function keys work as follow:

* The unshifted key sets the variable value
* The shifted value solves for the given variable
* The second-shifted value recalls the variable value

The menu also shows the current value in menu label.

Fixes: #925

Signed-off-by: Christophe de Dinechin <christophe@dinechin.org>