fonts: Add support for fixed-width digits
When the setting `FixedWidthDigits` is set, all digits have the same
width. Note that when the setting is changed, it may only come in
effect after the relevant font cache entries are flushed out. A good
way to achieve this objective is to display a text that uses a wide
variety of characters, like the output of the `VERSION` command.

Fixes: #938

Signed-off-by: Christophe de Dinechin <christophe@dinechin.org>