menus: Adjust size of menus to make descenders visible
Descenders in softkey menus were not really visible. This meant that a
unit like `kg` would be visually equivalent to `kq`.

This adjusts the size of menus by one pixel to make it visible, and
extends the clipping area by another pixel, which gives two extra
pixels when drawing in white on a black background.

This is not visually perfect, because as a result, menu entries
without a descender, e.g. 'mol' now appear to be a bit shifted up.

Fixes: #939

Signed-off-by: Christophe de Dinechin <christophe@dinechin.org>