parser: Accelerate and improve object parsing
Scan characters ahead of time in order to accelerate parsing.
This also solves various problem parsing units inside expressions, or
detecting error conditions such as `1.0+`.

While the functional improvements are numerous, the performance one is
just not there, for a reason that remains to be investigated:

Time to parse `DEMO.48S`, as measured by running `TEval` on
`« "/STATE/DEMO.48S" RCL »`:

Before: 358ms (DM32) 259ms (DM42)
 After: 427ms (DM32) 349ms (DM42)

Note that the main objective is not just performance, but also
improving battery life by reducing CPU utilization for a very common
operation. So it matters that we do not have a regression.

Fixes: #940

Signed-off-by: Christophe de Dinechin <christophe@dinechin.org>