solver: Replace `SolverPrecision` with `SolverImprecision`
The setting `SolverPrecision` was giving an absolute precision. It had
been set to 24 back when the default precision was 36 digits with the
IEEE-754 128-bit format.

When the number of digits was reduced to 24 after switching to the
variable-precision floating-point implementation, getting 24 digits
out of the solver became a game of chance, and unlikely for something
as simple as `sqrt(8)`.

Also add documentation for that setting.

Fixes: #947

Signed-off-by: Christophe de Dinechin <christophe@dinechin.org>