cycle: Update behaviour for several data types
The constants, equations and library items would give a type error
(`Bad argument type`), as originally reported in #957.

The symbol type would convert to text, which was inconsistent with
what happens for expressions, where we cycle graphical/text rendering.

The text type would convert to symbol, which could be used to create
invalid symbols that could not be parsed, making it potentially
dangerous.

The tag type would apply the `dtag` command, making it one case where
the cycle was not easily reversible.

All these now switched to the same behaviour we had for expression,
which is to cycle between graphical and text stack rendering. This is
almost guaranteed to be non-destructive.

Fixes: #957