put: Fix null-dereference checking the index
When putting an object in a list, we were testing the resulting index
correctly, but then dereferencing a null pointer before returning to
emit the error. On physical hardware, this leads to memory corruption
since null-dereferences are not correctly detected.

Also added some tests for `put` and `puti` that were missing.

Fixes: #954

Signed-off-by: Christophe de Dinechin <christophe@dinechin.org>