xroot: Reverse `xroot` arguments in algebraic parsing
The `xroot` command is a special case on HP50, in that `A B XROOT`
returns `XROOT(B;A)` and not `XROOT(A;B)` like for other commands.

Special-case the parsing and rendering of that command.

(This is super-ugly, all the more so that the HP50G advanced user
manual actually indicates the order depends on algebraic vs. RPN mode)

Fixes: #959

Reported-by: mahi2003@users.github.com
Signed-off-by: Christophe de Dinechin <christophe@dinechin.org>