wasm: Implement WASM support
This is an in-tree port of [Franco Trimboli's excellent work][wasm].

The goal is to be able to run the simulator in a browser.
This builds with emscripten, as installed by the `emsdk`.
The `emsdk` is added as a submodule, and initialized on first build.

To build the browser WASM version of the simulator, run `make wasm`.

Fixes: #993

Signed-off-by: Christophe de Dinechin <christophe@dinechin.org>
Suggested-by: Franco Trimboli <sunpazed@gmail.com>
Co-developed-by: Franco Trimboli <sunpazed@gmail.com>

[wasm]: https://github.com/sunpazed/db48x-wasm