complex: Give `Syntax error` for phasor without an angle
When entering a phasor without a phase, e.g. `1∡`, return a syntax
error instead of an internal error.

The internal error was a result of an undected error during parsing
(the outer shell code detects that no object was parsed but no error
reported, and reports an internal error when this happens).

Fixes: #989

Signed-off-by: Christophe de Dinechin <christophe@dinechin.org>