wasm: Run the RPL thread in a separate thread
Reserve the main thread for display updates and keyboard interaction,
like for the Qt version. This allows the web simulator to work
correctly when you run a program or when you access the firwmare
interface.

The downside is that you need to run that on a web server that is able
to set the following [HTTP headers][coop]:

```
Cross-Origin-Embedder-Policy: require-corp
Cross-Origin-Opener-Policy: same-origin
```

Unfortunately, it is not entirely clear if this can be made to work
easily with [GitHub pages][ghp].

Fixes: #994

Signed-off-by: Christophe de Dinechin <christophe@dinechin.org>

[coop]: https://web.dev/articles/coop-coep
[ghp]: https://github.com/orgs/community/discussions/13309