ui: Compute the correct refresh rate even when nothing is displayed
When we display nothing, we go through `ui.draw_header`, which may not
need to draw time or date. In that case, we were not correctly
computing the refresh time (i.e. calling `draw_refresh`.

If nothing else was drawn (e.g. no annunciator), then the refresh time
would remain super high, causing the display refresh to freeze.

A similar problem exited for the battery display, and was also fixed.
In the process, added the `BatteryRefresh` setting, defaulting to 3s,
to indicate how often we should check the battery voltage.

Fixes: #992

Signed-off-by: Christophe de Dinechin <christophe@dinechin.org>