ui: Create a non-empty dirty rectangle when clearing annunciators
When annunciators where cleared, the horizontal coordinate for the
dirty rectangle could end up empty, causing the next call to
`refresh_dirty` to do nothing.

The manifestation was that if clearing the shift indicator and nothing
else refreshed the header rows, the shift indicator would remain lit.

It's not entirely clear why the simulator behaved differently in that
scenario. It is likely due to the way timer interrupts are simulated,
which caused a faster refresh, making it almost impossible to see.

Fixes: #991

Signed-off-by: Christophe de Dinechin <christophe@dinechin.org>