units: Disable simplification during unit definition evaluation
When evaluating the definition of a unit, we disable simplifications,
which are not necessary if the unit was defined "correctly".

The evaluation of units as required for the equation simplification
loop was very expensive, because each run required going through the
`units.csv` file in order to check for units.

Fixes: #976

Signed-off-by: Christophe de Dinechin <christophe@dinechin.org>