parser: Parse degrees as an angle unit in phasors
Degrees can be either an angle marker in phasors, but also part of a
DMS notation. So parsing `10°` results in a DMS value for 10 degrees,
which is 10. As a result, the phasor parser was not seeing the degree
sign and interpreted the result as not having a unit.

Add a `separator` field to the `parser` class to record that we are
inside a phasor, so that the `integer` parser can avoid entering DMS
mode.

Fixes: #983

Signed-off-by: Christophe de Dinechin <christophe@dinechin.org>