menu: Insert `iferr-then-else` correctly from menu
In order to correctly insert the `iferr` statements, we need them to
be put outside of the range of immediates.

The rendering was also broken because `IfErrThenElse` is a two byte
opcode and the rendering code was using `payload(o)` (which optimizes
for the 1-byte opcodes) instead of `o->payload()` (which works for all
opcode sizes).

Added some tests, that triggered #1015 and #1016, to be dealt with.

Fixes: #1014

Signed-off-by: Christophe de Dinechin <christophe@dinechin.org>