condidionals: Avoid infinite loop evaluating condition
When the condition was a program or expression, the `conditional`
opcode was deferred while the program was evaluated.

That could lead to a situation where an expression would evaluate as
itself, and we would never finish the evaluation of the expression. An
example was evaluating `'FAIL'` that would evaluate as `'FAIL'` and
never make forward progress.

The implemented solution adds a `need_conditional` opcode that is
deferred to evaluate just before the next evaluation of the original
conditional, and demands that the conditional after evaluation be a
truth value. If after evaluating `'FAIL'` we get `'FAIL'`, this ends
up with a `Bad argument type` error message.

Fixes: #1016

Signed-off-by: Christophe de Dinechin <christophe@dinechin.org>