ui: Accept digits to select stack level in interactive stack
When entering digits in interactive stack mode, jump to the
corresponding level. If the stack is big enough, you can use sequences
of digits, e.g. type `1` `2` then `3` to go to level 123. If you type
`0` and the number becomes big, then it goes to the highest level of
the stack.

Fixes: #1029

Signed-off-by: Christophe de Dinechin <christophe@dinechin.org>