blitter: Adjust the right margin
The right margin only needs to be rounded to the nearest byte.

Using the scanline is wrong for the hardware screens, where the value
is artificially inflated (416 for 400 actual pixels) because DMCP uses
the extra two bytes for its own internal processing. Swapping
coordinates with 416 instead of 400 ends up messing up that data.

Fixes: #1053

Signed-off-by: Christophe de Dinechin <christophe@dinechin.org>