characters: Display correct content for built-in menus
When the `ShowBuiltinCharacters` setting is active, the index of the
built-in menus was off by one.

That's because if we find a match, we exit the loop without
incrementing the `menu` variable, so we find a match for the builtin
that follows as well.

A similar problem exists for constants.

Fixes: #1054

Signed-off-by: Christophe de Dinechin <christophe@dinechin.org>