solver: Make precision relative to equation sides magnitude
When we have an actual equation in the solver, e.g. `L=R`, then we can
make the epsilon of the solver relative to `L` or `R`.

This allows equations with large values for `L` and `R`, like the
Elastic Buckling examples given, to solve correctly (i.e. without a
`Sign reversal` error message) with the default settings.

Fixes: #1088

Signed-off-by: Christophe de Dinechin <christophe@dinechin.org>