solver: Do not change unit when storing in a solver variable
When storing in a solver variable, we need to check that units are
compatible, but we should not change the unit selected by the user.

For example, the Simple Deflection equations example in the HP50G
advanced reference manual are given in obsolete units. We want the
units to remain as is, since it's the user's preference.

Signed-off-by: Christophe de Dinechin <christophe@dinechin.org>