equations: Add missing equations from "Columns and Beams"
Add the missing equations from "Columns and Beams" section of the
HP50G advanced reference manual.

Notice that some of the descriptions are incorrect in the manual, and
that the correct equations had to be fetched from the HP50G itself.

Foe example, the `Simple Moment` example defines `c`, which is not set
in the documented equation. The actual equation contains `IFTE`
expressions, which require the fix for #1093 to work.

Signed-off-by: Christophe de Dinechin <christophe@dinechin.org>