equations: Add missing units in some equations
There is no unit for some of the equations in the Columns and Beams
category. Notably, some equations are missing the units for `A` and
`I. Adding them back to avoid `Inconsistent units` error when the
variables are not already give a value with a unit (which was the case
from the easlier tests in the test suite).

Fixes: #1113

Signed-off-by: Christophe de Dinechin <christophe@dinechin.org>