Implement `subst` and `where`
Implement the `subst` and `where` commands.

The `where` command can also be used as an operator `|` in
expressions, for example `'A+B|A=X-3'` will evaluate as `'X-3+B'`.

The HP syntax with parentheses for multiple substitutions, for example
`'A+B|(A=3;B=2)'`, is not implemented in this commit. It can be
replaced with sequences, like `'A+B|A=3|B=2'`.

Fixes: #1187

Signed-off-by: Christophe de Dinechin <christophe@dinechin.org>