arithmetic: Apply arithmetic and functions to equalities
When dealing with equalities, ensure that we apply arithmetic and
functions to both sides of the equation.

For example, `'A=B' 3 +` will result in `'A+3=B+3'`.

When both sides of an arithmetic operation are equalities, then apply
left to left and right to right, i.e. `'A=B' 'C=D' *` will return
`'A*C=B*D`'.

Fixes: #1198

Signed-off-by: Christophe de Dinechin <christophe@dinechin.org>