explode: Extract arguments for a function call into arguments
The `Obj→` command now explodes the argments to a function call into
an array.  It turns a function call object into its arguments.

Optimize the internal logic for grabbing arguments during expression
rewrites to avoid building expressions when matching a single object.

This exposed a bug in rewrite rules for `collect` and `expand`, which
were reordering terms using two different methods, causing the
rewrites to go in a loop.

In order to use only the reordering rule based on memory ordering,
while at the same time putting constants before variables (i.e.
rewrite `A*2` as `2*A`), the order of memory comparison when objects
have a different type is now reversed.

Signed-off-by: Christophe de Dinechin <christophe@dinechin.org>

Signed-off-by: Christophe de Dinechin <christophe@dinechin.org>