library: Add possibility to put individual objects in library
Add a new syntax in the library configuration, where an entry
that begins with `=` loads a file from disk.

If the entry is `=` without any name, then the name of the entry is
used as a file name. For example, `"KineticEnergy", "="` will load the
definition of `KineticEnergy` from the source file file named
`library/KineticEnergy.48s`.

A name can be specified as well. For example, an entry such as
`"Cst", "=/config/constants.csv"` will load the CSV file for
user-defined constants directly in memory.

This makes it possible to move all the demos as library entries.

This also fixes a bug where only the first category in external
constants file was properly loaded.

Fixes: #1210

Signed-off-by: Christophe de Dinechin <christophe@dinechin.org>