expressions: Do not read back vertical offset after `parentheses`
The `parentheses` function does not actually adjust the `g.voffset`
field, so reading it back after calling the function for the
right-hand side ends up reading back the one for the left hand side.
This lead to incorrect vertical adjustment computations.

The fix is to remove the readback of `g.voffset` to preserve the one
for the inner expression.

While at it, slightly adjust the rendering of `abs`.

Fixes: #1216

Signed-off-by: Christophe de Dinechin <christophe@dinechin.org>