units: Make sure that `ohm` can be spelled `Ω` or `Ω`
There is a dedicated character for Ohm, `Ω` (U8486). It is visually
identical to capital Greek Omega `Ω` (U937).

Initially, the choice was made to use 8486, since this is dedicated to
this particular unit. However, when Jean Wilson submitted his
equations, they failed with `Inconsistent units` errors, because he
had used the "regular" capital Omega. I reasoned at the time that not
using a dedicated character was consistent with what happens with
other units. For example, we don't use a special `A` for Ampere.

Unfortunately, there was a leftover `Ω` (U8436) in the definition of
alternate spelling `ohm`, and it turns out that `ohm` itself was used
in the definition of `H` (Henry), which is used in the unit for
`μ0` (`H/m`). So now we had two units that were spelled the same way,
and `1_H UBASE` would give `1_Ω*s`, which did not match the "Ohm" in
the definition of `F`.

The fixes the computation of `Ⓒε0 Ⓒμ0 * sqrt inv →Num UBASE`, which
now correctly gives a result in `m/s`.

Fixes: #1224

Signed-off-by: Christophe de Dinechin <christophe@dinechin.org>