decimal: Preserve trailing decimal in `FIX` mode
The mechanism to elminate the trailing decimal would incorrectly
suppress it in `FIX` mode.

We also need to make sure we never suppress the trailing decimal
while editing, otherwise we might change the data type from decimal to
integer, which is probably not a side effect we want from editing.

Tested that the rounding was correct with 500.0 and 500.5 and that the
trailing decimal does not show in either case when we use `0 FIX`, but
shows for higher values of `FIX`.

Fixes: #1236

Signed-off-by: Christophe de Dinechin <christophe@dinechin.org>