help: Parse help topics according to command parsing names
Make sure that we find the name topic irrespective of the spelling.

Replace help sections that contained something like `Duplicate (DUP)`
with a single spelling, e.g. `Duplicate`. Any spelling will do,
picking the default one is probably best.

This is a better approach that what we were doing earlier, some
complicated ad-hoc parsing with the expectation that the title would
look something like `IntegerPart (IP, IntPart)`. This was fragile,
hard to maintain, and not compatible with standard markdown.

Improve the rendering of the help file to match. Make the rendering of
help file titles denser, showing more on one screen, and show the name
of all other aliases on the right.

Adjust the rendering of keywords in `SUBTITLE` and `CODE` sections to
match the current preference (i.e. display `DUP` or `Duplicate`)

Fixes: #1231

Signed-off-by: Christophe de Dinechin <christophe@dinechin.org>