command: Switch to binary search for command parsing
Use the sorted spellings array built initially for the catalog while
parsing commands. Since there are currently 1182 distinct spellings in
the table as of today, this brings the number of iterations down by
about two orders of magnitude.

This optimization was made necessary due to the time required to bring
up help topics that are near the end of the help file. This is because
the recent change to adjust the help file to the current command
display style requires a fairly large number of command lookups.

For example, `→Unit` which is at line 10938 in the file takes about
* about 6s on the 0.8.0 version using the former ad-hoc scanner
* about 10s just prior to this optimization, and
* about 6s with this optimization.

This means that the command lookup no longer contributes a significant
amount to the help file scanning.

Signed-off-by: Christophe de Dinechin <christophe@dinechin.org>