help: Index the  help file for performance
Since loading data from the help file is so slow on the DM42, index
the file in order to be able to jump to the right location directly.

With this optimization, jumping to the last entry in the index file,
which is `→Unit`, takes less about 1s, as opposed to about 6s before.
Jumping to the last indexed entry (Volumic Density Magnetic Energy)
also takes about 1s.

Fixes: #1237

Signed-off-by: Christophe de Dinechin <christophe@dinechin.org>