ui: Clear selection when BSP is used with a non-empty selection
The expected behaviour of modern text editors when there is a
selection is that the BSP or Delete keys clear that selection.

The previous behaviour of DB48X, which was to clear the character at
left or right of cursor, is a bit surprising. Switch to a more usual
behaviour.

Signed-off-by: Christophe de Dinechin <christophe@dinechin.org>