atan: Adjust argument reduction for values in [0.5, 1.0]
Argument values in the range 0.5 to 1.0 were incorrectly reduced,
resulting in slow convergence and insufficient precision. This was the
case for a value like `0.74` which appears during the conversion of
`[ 1 2 3 ]` from rectangular to spherical, which resulted in an
incorrect conversion back to rectangular.

With the change, converting `[ 1 2 3 ]` to spherical and back and
subtracting `[ 1 2 3]` results in values in the order of `1E-23` in
radians as well as in degrees (specifically `[ -8E-23 -3E-23 4E-23]`.

Fixes: #1240

Signed-off-by: Christophe de Dinechin <christophe@dinechin.org>