graphics: Add missing font parameter to `→Grob`
The `→Grob` command, as documented in the HP50 ARM, takes two
arguments and not just one. The first stack level specifies the font
size. Size 0 is specified as "the same as in EquationWriter".
Interpret this on DB48X as being the font selected by `StackFont`.

Also add missing documentation for other graphing operations,
with examples.

Fixes: #1242

Signed-off-by: Christophe de Dinechin <christophe@dinechin.org>