settings: Avoid change in color settings after `Purge`
There was a bug in the `Purge` command for patterns, where we took
the default value from the `ids.tbl` file and not from the `settings`
constructor. As a result, the value after `ResetModes` and the value
after `Purge` were not identical. This showed up in some graphic tests
failing (different colors for some UI elements) depending on whether
the `settings_by_name` test had run before or not.

In order to solve this, the definitions of the initial settings have
been moved to `ids.tbl`, adding a few macros to help with color
pattern definitions in both color and black-and-white cases.

Signed-off-by: Christophe de Dinechin <christophe@dinechin.org>