equations: Application of `|` operator extracts value of equation
When a library equation is given as input, apply the `|` operator to
the value of the equation.

Additionally, recognize the variables with a unit in the equation, and
convert the input value to the target unit. For example, in the
Relativity Mass Energy equation, the variable `m` is given the `kg`
unit, so that applying `'ⒺRelativity Mass Energy|m=1_g'` will return
`'E=1/1000_kg*Ⓒc^2'`.

With these changes, it is possible to solve library equations
in a single operation with `Root`. For example:
`'Root(ⒺRelativityMassEnergy|m=(1_g);E;1_MJ)'` evaluates directly as
`:E:89 875 517.87368 17639 99999 9_MJ`.

Fixes: #1245

Signed-off-by: Christophe de Dinechin <christophe@dinechin.org>