root: Update the global variable after solving
The HP calculators update the variable value after solving
(and probably during solving too, but let's avoid that for performance
reason)

Implementing the corresponding behaviour also exposed a more important
problem during parsing of commands requiring a symbolic argument, such
as `Root`, because expression arguments were inserted "in line" and
therefore evaluated before evaluating the function.

For example, `'ROOT(X+3;X;1)'` would evaluate as the same program as
`X 3 + X 1 ROOT` and not `'X+3' 'X' 1 ROOT`.

Fixing that required adding some infrastructure to check if a function
has symbolic arguments, and if so, check if a given argument needs to
be preserved in symbolic form.

Fixes: #1251

Signed-off-by: Christophe de Dinechin <christophe@dinechin.org>