solver: Improve behaviour when a solution is not found
When there is no solution, attempt to converge more rapidly toward an
extremum.

After considering and experimenting with a more classical
[Kahan approach][k], I went for a somewhat simpler implementation that
works as follows:

Once you have found a range [lx, hx] with values [ly, hy] where lx is
where you have the lowest result so far, and hx is the second
lowest (so the interval [lx, hx] can be backwards), we compute the
secant x and corresponding y.

If y improves over ly, then we have found an absolute best so far.
lx becomes hx, ly becomes hy, x becomes lx and y becomes ly.

If y improves over hy, then we have found the second best, and we
move x to hx and y to hy.

If y does not improve over either, or we had an error computing y,
then we enter a degraded mode where we will explore other locations
than the secant. This is where we diverge from Kahan.

If we ever had a positive and negative y, then we simply bisect.
This will always converge, although it may fail to find a solution,
as if you try to find a solution to 1/x=0 between -1 and 1.

Otherwise, we move outwards if x is between lx and hx, inwards
if it is outside (the most common case). Moving inwards consist in
reducing the interval between hx and x by a factor of -e, moving
outwards consists in multiplying the interval between hx and x by a
factor of e. Since we move relative to hx and not lx, this means that
we will explore both the interval between hx and lx and the interval
outside of it, on both sides.

[k]: https://people.eecs.berkeley.edu/~wkahan/Math128/SOLVEkey.pdf

Signed-off-by: Christophe de Dinechin <christophe@dinechin.org>