ui: Avoid garbage at right of menu area
When we are refreshing lazily, avoid garbage at the right of menu.
This makes it possible to drop an extra forced refresh in the test
harness that has the nasty side effect of clearing the transient
object, which made the tests for transient object in the solver
incorrect: the `images/solver-eqdisplay.png` was supposed to show the
equation but wasn't because of the forced refresh.

Having a forced refresh was a bad idea anyway, since it may hide a
number of display refresh glitches.

Signed-off-by: Christophe de Dinechin <christophe@dinechin.org>