variables: When using `3▶A`, put the stored `3` on the stack
When we use `3▶A`, we created a persistent global copy of a possibly
temporary stack object. This means that we have two copies of `3`, one
on the stack and one in globals. Since it's on the stack, the copy on
the stack would remain in case of garbage collection.

Use the stored value for the stack instead, so that the value on the
stack can be garbage collected. This matters if the object in question
is big, for example if we are manipulating an array.

Fixes: #1265

Signed-off-by: Christophe de Dinechin <christophe@dinechin.org>